+++ webApiReact/ViewModels/ReportAnswersModel.cs
@@ -10,16 +10,16 @@ namespace webApiReact.ViewModels
 
         public required int K_PRED { get; set; }
 
-        public int? P0 { get; set; }
-        public int? P1 { get; set; }
-        public int? P2 { get; set; }
-        public int? p3 { get; set; }
-        public int? p4 { get; set; }
-        public int? p5 { get; set; }
-        public int? p6 { get; set; }
-        public int? p7 { get; set; }
-        public int? p8 { get; set; }
-        public int? p9 { get; set; }
-        public int? p10 { get; set; }
+        public long? P0 { get; set; }
+        public long? P1 { get; set; }
+        public long? P2 { get; set; }
+        public long? p3 { get; set; }
+        public long? p4 { get; set; }
+        public long? p5 { get; set; }
+        public long? p6 { get; set; }
+        public long? p7 { get; set; }
+        public long? p8 { get; set; }
+        public long? p9 { get; set; }
+        public long? p10 { get; set; }
     }
 }
